+++ include/transformer/transformer.hpp
@@ -47,7 +47,7 @@ class TransformerNetwork : public Network {
         const mlx::core::array& backward(const vector<int>& y_true, const mlx::core::array& y_pred);
         void infer(const vector<vector<int>>& encoder_token_ids, BPETokenizer* tokenizer, const string& csv_path) const;
         void infer_live(BPETokenizer* tokenizer) const;
-        double compute_validation_loss(vector<vector<int>>& encoder_tokens_val, vector<vector<int>>& decoder_tokens_val);
+        float compute_validation_loss(vector<vector<int>>& encoder_tokens_val, vector<vector<int>>& decoder_tokens_val);
         void reset_gradients();
         void normalize_gradients(int batch_size);
         void train(
